requirements.txt
opencv-python
numpy
mediapipe
fastapi
uvicorn
python-socketio
python-multipart
jinja2
